Move all CUDA code to the plugins/ directory
Keep RiemannFit.h in the interface, as it is include-only.
