+++ RecoPixelVertexing/PixelTrackFitting/plugins/BuildFile.xml
@@ -1,4 +1,5 @@
-<use   name="RecoPixelVertexing/PixelTrackFitting"/>
-<library   file="*.cc" name="RecoPixelVertexingPixelTrackFittingPlugins">
-  <flags   EDM_PLUGIN="1"/>
+<use name="cuda"/>
+<use name="RecoPixelVertexing/PixelTrackFitting"/>
+<library file="*.cc *.cu" name="RecoPixelVertexingPixelTrackFittingPlugins">
+  <flags EDM_PLUGIN="1"/>
 </library>
